Implement the visit and accept method generation from 5.3.3
This needs some slight tweaks to work around the Go limitation that
type parameters can't be added in methods. Since we don't want to
parameterise the ast itself by the the return type of some visitor, we
define an Acceptor[R] type for each expr type, and a conversion
function asAcceptor to turn each expr into the corresponding
acceptor. Then Accept can be a generic top-level function rather than
a method in Expr.